src/requirements.txt
certifi==2019.11.28
cffi==1.14.0
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
kiwisolver==1.1.0
matplotlib==3.2.0rc1
mkl-fft
mkl-random
mkl-service==2.3.0
networkx==2.4
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
Pillow==6.2.2
progress==1.5
progressbar==2.5
pycocotools==2.0
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.45.0
